+++ experimental/client-go/bindings/core/wandbcore.go
@@ -12,12 +12,12 @@ import "C"
 import (
 	"unsafe"
 
-	"github.com/wandb/wandb/core/pkg/gowandb"
-	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
-	"github.com/wandb/wandb/core/pkg/gowandb/opts/sessionopts"
-	"github.com/wandb/wandb/core/pkg/gowandb/runconfig"
 	"github.com/wandb/wandb/core/pkg/service"
-	"github.com/wandb/wandb/gowandb/internal/gowandb/internal_runopts"
+	"github.com/wandb/wandb/experimental/client-go/gowandb"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/internal/gowandb/internal_runopts"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/opts/runopts"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/opts/sessionopts"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/runconfig"
 )
 
 // globals to keep track of the wandb session and any runs
