+++ experimental/client-go/go.mod
@@ -1,3 +1,8 @@
 module github.com/wandb/wandb/experimental/client-go/gowandb
 
+replace (
+	// github.com/wandb/wandb => /Users/dimaduev/dev/sdk
+	github.com/wandb/wandb/experimental/client-go/gowandb => /Users/dimaduev/dev/sdk/experimental/client-go
+)
+
 go 1.22.2
