+++ experimental/client-go/pkg/gowandb/gowandb.go
@@ -2,7 +2,7 @@
 package gowandb
 
 import (
-	"github.com/wandb/wandb/gowandb/opts/sessionopts"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/opts/sessionopts"
 )
 
 type History map[string]interface{}
