+++ experimental/client-go/pkg/gowandb/run.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
 	"github.com/wandb/wandb/core/pkg/service"
-	"github.com/wandb/wandb/gowandb/runconfig"
+	"github.com/wandb/wandb/experimental/client-go/gowandb/runconfig"
 )
 
 type Settings map[string]interface{}
